fix: address review — dedupe fmtTokens, fix zero-token handling
- Move fmtTokens to core.js shared util, remove duplicates from
  dashboard.js and logs.js
- Change v > 0 to v >= 0 in PassthroughStreamProcessor._extract_usage
  so zero-token responses are distinguished from absent data
- Add test for zero prompt_tokens edge case
