+++ src/llm_rosetta/gateway/admin/js/dashboard.js
@@ -566,13 +566,6 @@ function renderPersistence(p, totalReq) {
   `;
 }
 
-function fmtTokens(n) {
-  if (n >= 1e9) return (n/1e9).toFixed(1) + 'B';
-  if (n >= 1e6) return (n/1e6).toFixed(1) + 'M';
-  if (n >= 1e3) return (n/1e3).toFixed(1) + 'K';
-  return String(n);
-}
-
 function renderStats(d) {
   const uptime = formatDuration(d.uptime_seconds);
   const errRate = (d.error_rate * 100).toFixed(1) + '%';
