+++ src/llm_rosetta/gateway/admin/js/logs.js
@@ -33,13 +33,6 @@ function resolveProviderName(apiType) {
   return apiType;
 }
 
-function fmtTokens(n) {
-  if (n >= 1e9) return (n/1e9).toFixed(1) + 'B';
-  if (n >= 1e6) return (n/1e6).toFixed(1) + 'M';
-  if (n >= 1e3) return (n/1e3).toFixed(1) + 'K';
-  return String(n);
-}
-
 function renderLogs(entries, total) {
   const tbody = document.getElementById('logTable');
   if (entries.length === 0) {
